new VFS/tmpfs operations, syscalls, final touches for 0.2.3

+++ src/com/fs/vfs.c
@@ -22,6 +22,7 @@
 #include <lib/printf.h>
 #include <lib/util.h>
 #include <stdbool.h>
+#include <stdlib.h>
 
 #define GET_VLINK_CHILD(vn)          \
     while (NULL != vn->vlink.next) { \
@@ -50,11 +51,11 @@ void com_fs_vfs_vlink_set(com_vnode_t *parent, com_vnode_t *vlink) {
 int com_fs_vfs_close(com_vnode_t *vnode) {
     GET_VLINK_CHILD(vnode);
 
-    if (NULL != vnode->ops->close) {
-        return vnode->ops->close(vnode);
+    if (NULL == vnode->ops->close) {
+        return ENOSYS;
     }
 
-    return ENOTSUP;
+    return vnode->ops->close(vnode);
 }
 
 int com_fs_vfs_lookup(com_vnode_t **out,
@@ -104,12 +105,16 @@ int com_fs_vfs_lookup(com_vnode_t **out,
 
             com_vnode_t *dir = ret;
             ret              = NULL;
-            dir->ops->lookup(&ret, dir, path, section_end - path);
+            int fsret = dir->ops->lookup(&ret, dir, path, section_end - path);
 
             if (NULL == ret) {
                 COM_FS_VFS_VNODE_RELEASE(dir);
                 *out = NULL;
                 return ENOENT;
+            } else if (0 != fsret) {
+                COM_FS_VFS_VNODE_RELEASE(dir);
+                *out = NULL;
+                return fsret;
             }
 
             while (NULL != ret->mountpointof) {
@@ -137,6 +142,10 @@ int com_fs_vfs_create(com_vnode_t **out,
                       const char   *name,
                       size_t        namelen,
                       uintmax_t     attr) {
+    if (NULL == dir->ops->create) {
+        return ENOSYS;
+    }
+
     return dir->ops->create(out, dir, name, namelen, attr);
 }
 
@@ -145,6 +154,10 @@ int com_fs_vfs_mkdir(com_vnode_t **out,
                      const char   *name,
                      size_t        namelen,
                      uintmax_t     attr) {
+    if (NULL == parent->ops->mkdir) {
+        return ENOSYS;
+    }
+
     return parent->ops->mkdir(out, parent, name, namelen, attr);
 }
 
@@ -152,6 +165,10 @@ int com_fs_vfs_link(com_vnode_t *dir,
                     const char  *dstname,
                     size_t       dstnamelen,
                     com_vnode_t *src) {
+    if (NULL == src->ops->link) {
+        return ENOSYS;
+    }
+
     if (0 == src->ops->link(dir, dstname, dstnamelen, src)) {
         com_vnode_t *link = NULL;
 
@@ -164,11 +181,14 @@ int com_fs_vfs_link(com_vnode_t *dir,
         }
     }
 
-    // TODO: improve error handling here
     return ENOENT;
 }
 
 int com_fs_vfs_unlink(com_vnode_t *dir, const char *name, size_t namelen) {
+    if (NULL == dir->ops->unlink) {
+        return ENOSYS;
+    }
+
     return dir->ops->unlink(dir, name, namelen);
 }
 
@@ -179,6 +199,10 @@ int com_fs_vfs_read(void        *buf,
                     uintmax_t    off,
                     uintmax_t    flags) {
     GET_VLINK_CHILD(node);
+
+    if (NULL == node->ops->read) {
+        return ENOSYS;
+    }
 
     size_t dump = 0;
     if (NULL == bytes_read) {
@@ -197,6 +221,10 @@ int com_fs_vfs_write(size_t      *bytes_written,
                      uintmax_t    flags) {
     GET_VLINK_CHILD(node);
 
+    if (NULL == node->ops->write) {
+        return ENOSYS;
+    }
+
     size_t dump = 0;
     if (NULL == bytes_written) {
         bytes_written = &dump;
@@ -207,32 +235,67 @@ int com_fs_vfs_write(size_t      *bytes_written,
 }
 
 int com_fs_vfs_resolve(const char **path, size_t *pathlen, com_vnode_t *link) {
+    if (NULL == link->ops->resolve) {
+        return ENOSYS;
+    }
+
     return link->ops->resolve(path, pathlen, link);
 }
 
 int com_fs_vfs_readdir(void        *buf,
-                       size_t      *buflen,
+                       size_t       buflen,
+                       size_t      *bytes_read,
                        com_vnode_t *dir,
                        uintmax_t    off) {
-    return dir->ops->readdir(buf, buflen, dir, off);
+    if (NULL == dir->ops->readdir) {
+        return ENOSYS;
+    }
+
+    return dir->ops->readdir(buf, buflen, bytes_read, dir, off);
 }
 
 int com_fs_vfs_ioctl(com_vnode_t *node, uintmax_t op, void *buf) {
     GET_VLINK_CHILD(node);
+
+    if (NULL == node->ops->ioctl) {
+        return ENOSYS;
+    }
+
     return node->ops->ioctl(node, op, buf);
 }
 
 int com_fs_vfs_isatty(com_vnode_t *node) {
     GET_VLINK_CHILD(node);
+
+    if (NULL == node->ops->isatty) {
+        return ENOSYS;
+    }
+
     return node->ops->isatty(node);
 }
 
 int com_fs_vfs_stat(struct stat *out, com_vnode_t *node) {
+    if (NULL == node->ops->stat) {
+        return ENOSYS;
+    }
+
     // TODO: use vlink, but needs to me masked
     kmemset(out, sizeof(struct stat), 0);
     return node->ops->stat(out, node);
 }
 
 int com_fs_vfs_truncate(com_vnode_t *node, size_t size) {
+    if (NULL == node->ops->truncate) {
+        return ENOSYS;
+    }
+
     return node->ops->truncate(node, size);
+}
+
+int com_fs_vfs_vnctl(com_vnode_t *node, uintmax_t op, void *buf) {
+    if (NULL == node->ops->vnctl) {
+        return ENOSYS;
+    }
+
+    return node->ops->vnctl(node, op, buf);
 }
